+++ src/components/Typography/Typography.tsx
@@ -8,13 +8,15 @@ import styles from "./Typography.module.scss";
 
 type TypographyH1Props = ComponentProps<"h1"> & {
   mt?: string;
+  mb?: string;
 };
 
 const TypographyH1: FC<TypographyH1Props> = (props) => {
-  const { children, className, style, ...rest } = props;
+  const { children, className, style, mt, mb, ...rest } = props;
 
   const computedStyle = {
-    marginTop: props.mt,
+    marginTop: mt,
+    marginBottom: mb,
     ...style,
   };
 
@@ -34,16 +36,18 @@ const TypographyH1: FC<TypographyH1Props> = (props) => {
 
 type TypographyH2Props = ComponentProps<"h2"> & {
   mt?: string;
+  mb?: string;
 };
 
 const TypographyH2: FC<TypographyH2Props> = (props) => {
-  const { children, className, style, ...rest } = props;
+  const { children, className, style, mt, mb, ...rest } = props;
 
   const theme = useTheme();
 
   const computedStyle = {
     "--typography-h2-border-color": theme.colors.gray["200"],
-    marginTop: props.mt,
+    marginTop: mt,
+    marginBottom: mb,
     ...style,
   };
 
@@ -63,13 +67,15 @@ const TypographyH2: FC<TypographyH2Props> = (props) => {
 
 type TypographyH3Props = ComponentProps<"h3"> & {
   mt?: string;
+  mb?: string;
 };
 
 const TypographyH3: FC<TypographyH3Props> = (props) => {
-  const { children, className, style, ...rest } = props;
+  const { children, className, style, mt, mb, ...rest } = props;
 
   const computedStyle = {
-    marginTop: props.mt,
+    marginTop: mt,
+    marginBottom: mb,
     ...style,
   };
 
@@ -84,18 +90,48 @@ const TypographyH3: FC<TypographyH3Props> = (props) => {
   );
 };
 
+// Typography.H4
+// =============================================================================
+
+type TypographyH4Props = ComponentProps<"h4"> & {
+  mt?: string;
+  mb?: string;
+};
+
+const TypographyH4: FC<TypographyH4Props> = (props) => {
+  const { children, className, style, mt, mb, ...rest } = props;
+
+  const computedStyle = {
+    marginTop: mt,
+    marginBottom: mb,
+    ...style,
+  };
+
+  return (
+    <h4
+      className={classNames(styles.h4, className)}
+      style={computedStyle}
+      {...rest}
+    >
+      {children}
+    </h4>
+  );
+};
+
 // Typography.P
 // =============================================================================
 
 type TypographyPProps = ComponentProps<"p"> & {
   mt?: string;
+  mb?: string;
 };
 
 const TypographyP: FC<TypographyPProps> = (props) => {
-  const { children, className, style, ...rest } = props;
+  const { children, className, style, mt, mb, ...rest } = props;
 
   const computedStyle = {
-    marginTop: props.mt,
+    marginTop: mt,
+    marginBottom: mb,
     ...style,
   };
 
@@ -117,6 +153,7 @@ type TypographyType = {
   H1: typeof TypographyH1;
   H2: typeof TypographyH2;
   H3: typeof TypographyH3;
+  H4: typeof TypographyH4;
   P: typeof TypographyP;
 };
 
@@ -124,6 +161,7 @@ const Typography: TypographyType = {
   H1: TypographyH1,
   H2: TypographyH2,
   H3: TypographyH3,
+  H4: TypographyH4,
   P: TypographyP,
 };