+++ src/components/Typography/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-10-03: Added Typography.H2, Typography.H3 and Typography.P components.
   Support the `mt` prop for margin-top spacing & update CSS styles to make it
   look better.
+- 2025-10-04: Add Typography.H4 component. Support `mb` prop for margin-bottom
+  spacing. Increase H1 font size and improve prop handling consistency.